Add internationalization and improve project structure
- Add i18next and react-i18next for internationalization
- Move metadata to a separate file in the app directory
- Create a 404 Not Found page
- Refactor dashboard page to use a separate DashboardContent component
- Improve error handling in healthcheck API
- Remove client-side metadata from layout
