+++ app/layout.tsx
@@ -1,4 +1,5 @@
-import type { Metadata } from "next";
+'use client';
+
 import { Geist, Geist_Mono } from "next/font/google";
 import { UserProvider } from '@auth0/nextjs-auth0/client';
 import Navbar from "@/components/Navbar";
@@ -15,11 +16,6 @@ const geistMono = Geist_Mono({
   variable: "--font-geist-mono",
   subsets: ["latin"],
 });
-
-export const metadata: Metadata = {
-  title: "assistant-ui App",
-  description: "Generated by create-assistant-ui",
-};
 
 export default function RootLayout({
   children,
