+++ app/not-found.tsx
@@ -1,0 +1,19 @@
+'use client';
+
+import Link from 'next/link';
+
+export default function NotFound() {
+    return (
+        <div className="flex flex-col items-center justify-center min-h-[60vh]">
+            <h1 className="text-4xl font-bold mb-4">404</h1>
+            <h2 className="text-xl mb-4">Page Not Found</h2>
+            <p className="text-gray-600 mb-8">The page you're looking for doesn't exist or has been moved.</p>
+            <Link
+                href="/"
+                className="px-4 py-2 bg-blue-600 text-white rounded-md hover:bg-blue-500 transition-colors"
+            >
+                Return Home
+            </Link>
+        </div>
+    );
+}
