+++ app/protected/dashboard/page.tsx
@@ -1,23 +1,6 @@
-'use client';
-
 import { getSession } from '@auth0/nextjs-auth0';
 import { redirect } from 'next/navigation';
-import { useTranslation } from 'react-i18next';
-
-function DashboardContent({ user }) {
-    const { t } = useTranslation(undefined, {
-        useSuspense: false
-    });
-
-    return (
-        <div className="space-y-6">
-            <h1 className="text-2xl font-bold">{t('protected_dashboard')}</h1>
-            <div>
-                <p>{t('welcome')}, {user.name}!</p>
-            </div>
-        </div>
-    );
-}
+import DashboardContent from './DashboardContent';
 
 export default async function DashboardPage() {
     const session = await getSession();
@@ -26,5 +9,5 @@ export default async function DashboardPage() {
         redirect('/api/auth/login');
     }
 
-    return <DashboardContent user={session.user} />;
+    return <DashboardContent user={{ name: session.user.name }} />;
 }
