+++ package.json
@@ -18,13 +18,13 @@
     "@auth0/nextjs-auth0": "^3.5.0",
     "@langchain/langgraph-sdk": "^0.0.36",
     "@testing-library/dom": "^10.4.0",
+    "i18next": "^24.2.2",
     "next": "^15.1.6",
     "react": "19.0.0",
     "react-dom": "19.0.0",
     "react-i18next": "^15.4.0",
     "tailwindcss-animate": "^1.0.7",
-    "zod": "^3.24.1",
-    "actions/upload-artifact": "v3"
+    "zod": "^3.24.1"
   },
   "devDependencies": {
     "@eslint/eslintrc": "^3.2.0",
